Move some configuration bits and files to new 'config' dir
