+++ config/configDesktopDev.js
@@ -1,4 +1,4 @@
-const utils = require( './utils' );
+const utils = require( '../utils' );
 const configDesktop = require( './configDesktop.js' );
 
 /**
